Make plugin popups open based on where plugin is located
